Adding backup strategy to migration options
It is now possible to backup the entire
database before a migration.
