+++ src/Hangfire.Mongo/Database/HangfireDbContext.cs
@@ -11,6 +11,8 @@ namespace Hangfire.Mongo.Database
     {
         private readonly string _prefix;
 
+        internal MongoClient Client { get; }
+
         internal IMongoDatabase Database { get; }
 
         /// <summary>
@@ -23,9 +25,9 @@ namespace Hangfire.Mongo.Database
         {
             _prefix = prefix;
 
-            var client = new MongoClient(connectionString);
+            Client = new MongoClient(connectionString);
 
-            Database = client.GetDatabase(databaseName);
+            Database = Client.GetDatabase(databaseName);
 
             ConnectionId = Guid.NewGuid().ToString();
         }
