+++ src/Hangfire.Mongo/MongoMigrationOptions.cs
@@ -1,8 +1,10 @@
-﻿namespace Hangfire.Mongo
+﻿using System;
+
+namespace Hangfire.Mongo
 {
 
     /// <summary>
-    /// The differnet supported migration strategies.
+    /// The supported migration strategies.
     /// </summary>
     public enum MongoMigrationStrategy
     {
@@ -20,6 +22,33 @@
         /// Migrate from old schema to new one.
         /// </summary>
         Migrate
+    }
+
+
+    /// <summary>
+    /// The supported backup strategies during migration.
+    /// </summary>
+    public enum MongoBackupStrategy
+    {
+        /// <summary>
+        /// No backup is made before migration.
+        /// </summary>
+        None,
+
+        /// <summary>
+        /// A collection-by-collection backup is made to the same database.
+        /// Recommended to use this if you store the Hangfire jobs in the 
+        /// same database as your actual application data.
+        /// This is the prefered and default backup strategy.
+        /// </summary>
+        Collection,
+
+        /// <summary>
+        /// Will copy the entire database into a new databse.
+        /// When using this backup strategy, access to the
+        /// "admin" database is required.
+        /// </summary>
+        Database
     }
 
 
@@ -45,7 +74,7 @@
         public MongoMigrationOptions(MongoMigrationStrategy strategy)
         {
             Strategy = strategy;
-            Backup = true;
+            BackupStrategy = MongoBackupStrategy.Collection;
             BackupPostfix = "migrationbackup";
         }
 
@@ -60,7 +89,24 @@
         /// If true, a backup of all Hangfire.Mongo collection will be
         /// performed before any migration starts.
         /// </summary>
-        public bool Backup { get; set; }
+        [Obsolete("Please use '" + nameof(BackupStrategy) + "' instead. This will be removed in next release.")]
+        public bool Backup
+        {
+            get => BackupStrategy != MongoBackupStrategy.None;
+            set
+            {
+                if (value != Backup)
+                {
+                    BackupStrategy = value ? MongoBackupStrategy.Collection : MongoBackupStrategy.None;
+                }
+            }
+        }
+
+
+        /// <summary>
+        /// The backup strategy to use before migrating.
+        /// </summary>
+        public MongoBackupStrategy BackupStrategy { get; set; }
 
 
         /// <summary>
@@ -69,6 +115,8 @@
         /// <remarks>
         /// The format for the backed up collection name is:
         /// {collection-name}.{schema-version}.{BackupPostfix}
+        /// The format for the backed up database name is:
+        /// {database-name}-{schema-version}-{BackupPostfix}
         /// </remarks>
         public string BackupPostfix { get; set; }
     }
